requirements.txt
Flask
pymongo
bleach
PyJWT